requirements.txt
alembic==1.3.1
appdirs==1.4.4
argcomplete==1.10.0
argh==0.26.1
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
autobahn==19.10.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.1.7
Beaker==1.10.0
beautifulsoup4==4.9.0
blinker==1.4
blivet==3.2.2
blivet-gui==2.1.13
Brlapi==0.7.0
cffi==1.14.0
cftime==1.0.4.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.1
conda==4.8.2
conda-package-handling==1.6.0
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
Cython==0.29.14
cytoolz==0.10.1
dasbus==0.2
dbus-python==1.2.16
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distro==1.4.0
Django==3.0.6
django-adminlte2==0.4.1
django-bootstrap4==1.1.1
django-import-export==2.1.0
django-localflavor==3.0.1
django-pagseguro==1.4.2
django-stdimage==5.1.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
fluidity-sm==0.2.0
fros==1.1
frozendict==1.2
gpg==1.13.1
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
invoke==1.4.1
itsdangerous==0.24
jdcal==1.4.1
Jinja2==2.11.2
langtable==0.0.51
ldap3==2.6.1
lexicon==1.0.0
libarchive-c==2.9
libcomps==0.1.15
lxml==4.4.1
lz4==3.0.2
Mako==1.1.1.dev0
MarkupPy==1.14
MarkupSafe==1.1.1
mod-wsgi==4.6.8
msgpack==0.6.2
mysqlclient==1.4.6
netCDF4==1.5.3
nftables==0.1
ntplib==0.3.3
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
ordered-set==3.1
pandas==1.0.4
paramiko==2.7.1
Paste==3.2.4
pid==2.2.3
Pillow==7.0.0
ply==3.11
productmd==1.26
progressbar2==3.51.3
psycopg2==2.8.5
pwquality==1.4.2
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.18.2
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.5
pydbus==0.6.0
pyenchant==2.0.0
Pygments==2.4.2
PyGObject==3.36.1
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pykickstart==3.24
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparted==3.11.4
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.0
python-editor==1.0.4
python-meh==0.48
python-snappy==0.5.4
python-stdnum==1.13
python-utils==2.4.0
python3-openid==3.1.0
pytz==2020.1
pyudev==0.22.0
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rpm==4.15.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.0
sepolicy==3.0
service-identity==18.1.0
setools==4.3.0
simplejson==3.17.0
simpleline==1.6
six==1.14.0
slip==0.6.4
slip.dbus==0.6.4
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sos==3.9
soupsieve==1.9.2
speaklater==1.3
SQLAlchemy==1.3.17
sqlparse==0.3.1
static3==0.7.0
systemd-python==234
tablib==2.0.0
Tempita==0.5.1
toolz==0.10.0
tqdm==4.45.0
Twisted==19.10.0
txaio==18.8.1
ujson==2.0
urllib3==1.25.7
virtualenv==20.0.21
Werkzeug==0.16.0
wsaccel==0.6.2
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
zope.event==4.2.0
zope.interface==4.7.1
